Unity 2023.2.0b8 C# reference source code

+++ Modules/TextCoreTextEngine/Managed/TextAssets/FontAsset.cs
@@ -1265,7 +1265,7 @@ namespace UnityEngine.TextCore.Text
         /// <returns></returns>
         public bool HasCharacter(char character, bool searchFallbacks = false, bool tryAddCharacter = false)
         {
-            return HasCharacter(character, searchFallbacks, tryAddCharacter);
+            return HasCharacter((uint)character, searchFallbacks, tryAddCharacter);
         }
 
         /// <summary>
